+++ src/checkbox/__test__/index.js
@@ -1,0 +1,15 @@
+Component({
+  data: {
+    value: ['1'],
+    single: false,
+    max: 2,
+  },
+
+  methods: {
+    handleChange(e) {
+      this.setData({
+        value: e.detail.value,
+      });
+    },
+  },
+});