+++ src/dialog/_example/base/index.js
@@ -1,0 +1,20 @@
+Component({
+  data: {
+    dialogKey: '',
+    showText: false,
+    showMultiText: false,
+    showTextAndTitle: false,
+    showMultiTextAndTitle: false,
+  },
+  methods: {
+    showDialog(e) {
+      const { key } = e.currentTarget.dataset;
+      this.setData({ [key]: true, dialogKey: key });
+    },
+
+    closeDialog() {
+      const { dialogKey } = this.data;
+      this.setData({ [dialogKey]: false });
+    },
+  },
+});