+++ src/radio/__test__/demo.test.js
@@ -1,16 +1,34 @@
-import simulate from 'miniprogram-simulate';
-import getDemoPath from '../../../test/utils/getDemoPath';
+/**
+ * 该文件为由脚本 `npm run test:demo` 自动生成，如需修改，执行脚本命令即可。请勿手写直接修改，否则会被覆盖
+ */
 
-describe('radio basic demo', () => {
-  let id;
-  beforeAll(() => {
-    id = simulate.load(getDemoPath('radio', 'basic'), {
-      less: true,
+import simulate from 'miniprogram-simulate';
+import path from 'path';
+
+const mapper = [
+  'base',
+  'controll',
+  'group',
+  'left-hook-shape',
+  'left-round',
+  'options',
+  'right-hook-shape',
+  'right-round',
+  'size',
+  'special',
+  'status',
+];
+
+describe('Radio', () => {
+  mapper.forEach((demoName) => {
+    it(`Radio ${demoName} demo works fine`, () => {
+      const id = simulate.load(path.resolve(__dirname, `../../radio/_example/${demoName}/index`), demoName, {
+        less: true,
+        rootPath: path.resolve(__dirname, '../..'),
+      });
+      const container = simulate.render(id);
+      container.attach(document.createElement('parent-wrapper'));
+      expect(container.toJSON()).toMatchSnapshot();
     });
-  });
-  test('demo should render correctly', async () => {
-    const container = simulate.render(id);
-    container.attach(document.createElement('parent-wrapper'));
-    expect(container.toJSON()).toMatchSnapshot();
   });
 });