+++ src/radio/_example/special/index.js
@@ -1,0 +1,6 @@
+Component({
+  data: {
+    activeImage: 'https://oteam-tdesign-1258344706.cos.ap-guangzhou.myqcloud.com/miniprogram/checkbox-checked.png',
+    inActiveImage: 'https://oteam-tdesign-1258344706.cos.ap-guangzhou.myqcloud.com/miniprogram/checkbox.png',
+  },
+});