+++ src/rate/_example/show-text/index.js
@@ -1,4 +1,4 @@
-Page({
+Component({
   data: {
     value: [3, 3, 3],
     texts: ['1分', '2分', '3分', '4分', '5分'],