+++ src/search/_example/status/index.js
@@ -1,7 +1,7 @@
 const placeholder = '搜索预设文案';
 const actionText = '取消';
 
-Page({
+Component({
   data: {
     searchBoxGroup: [
       {