+++ src/sticky/_example/sticky.ts
@@ -1,11 +1,1 @@
-Page({
-  data: {
-    container: null,
-  },
-  onLoad() {
-    this.setData({
-      // @ts-ignore
-      container: () => wx.createSelectorQuery().select('.box-c'),
-    });
-  },
-});
+Page({});