+++ src/tabs/_example/base/index.js
@@ -1,4 +1,9 @@
 Component({
+  data: {
+    stickyProps: {
+      zIndex: 2,
+    },
+  },
   methods: {
     onTabsChange(event) {
       console.log(`Change tab, tab-panel value is ${event.detail.value}.`);
@@ -6,6 +11,10 @@ Component({
 
     onTabsClick(event) {
       console.log(`Click tab, tab-panel value is ${event.detail.value}.`);
+    },
+
+    onStickyScroll(event) {
+      console.log(event.detail);
     },
   },
 });